Change input to return f64 from get_user_input

+++ src/user_input.rs
@@ -17,7 +17,7 @@ pub fn get_user_input(prompt: &str) -> Result<f64, Box<dyn Error>> {
     Ok(number)
 }
 
-pub fn input(prompt: &str) -> i32 {
+pub fn input(prompt: &str) -> f64 {
     loop {
         match get_user_input(prompt) {
             Ok(count) => {
